✨ feat: 이모지 코드 추가 및 타입 선언
