+++ src/@types/type.ts
@@ -19,3 +19,11 @@ export interface AuthFormProps {
   generation: string;
   image: string;
 }
+
+export type EmojiCode = 'HEART' | 'THUMBSUP' | 'LAUGH' | 'SAD' | 'CHECK' | 'ME';
+
+export interface EmojiType {
+  emojiCode: EmojiCode;
+  emojiCount: number;
+  isClicked: boolean;
+}
